Add Ordini

+++ src/model/Libro.java
@@ -28,7 +28,7 @@ public class Libro {
 						case 2: this.year = Integer.parseInt(fields[i]);
 						case 3: this.editor = fields [i];
 						case 4: this.isbn = fields [i];
-						case 5: this.price = Integer.parseInt(fields[i]);
+						case 5: this.price = Double.parseDouble(fields[i]);
 						case 6: this.copy = Integer.parseInt(fields[i]);
 					}
